FAQ Items (#4)
This PR introduces the first component for consumption: The FAQ Item. As
part of this PR, some work was done:

- To configure Tailwind, and adopt it as a peerDep
- Add install instructions to the README
- Update the storybook
- Update the file structure of dist

As part of this ticket, it has been revealed the potential need to
refactor and improve the props interface. Instead of passing tailwind
class names through props, we may want to go the direction of Tailwind
component classes. This also opens the door for theming in the future.

For now, this milestone allows us to make the package available for
consumption.